requirements.txt
starlette
tavily-python
uvicorn
asyncio
python-dotenv
mcp
google-generativeai